Add subcommand for non-draft releases
